Add initial instructions to set up dns and https for sync-endpoint
This commit takes a first pass and adding instructions for how to
setup a dns entry and enable HTTPS for deployments of the sync
endpoint. Right now it does not list detailed dns instructions since
I'm not sure if we want to recommend a specific provider. This commit
also captures changes made to the bringup process that automate the
provisioning of TLS certificates.